requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
boto3==1.14.58
botocore==1.17.58
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.1.0
djangorestframework==3.11.0
docutils==0.15.2
idna==2.10
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.5.2
pylint-django==2.0.15
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1